Update #9
I did:

++made 3 different enemy game objects:
One enemy attacks with a straight shot of a bullet.
the other attacks by targeting at the player's position and then
releases a bullet. and the last one attacks by colliding itself to the
player.

-- however, I do need help  with the last enemy programming to appear at
right side of screen and then follow up by colliding to player.

++I made replicated programming assets for all enemies.

++Organized my programming scripts into a folder.

++ and made each enemy a prefab with their own bullet attacks and bullet
guns as gameobjects.

end of Update #9.