Update #10
I suffered some really messed up game programming this passed week and i
was able to fix the problem by preventing the player from being
destroyed whenever the player is beginning to spawn.

I will configure a boss programming and fix the sound effect
programming.

Still configuring  the spawning control odf all enemies but i will need
some suggestions on how to stop an enemy in a specific amount of time to
cancel invoking of spawning functions.

and i will a special effect of spawning stars that will play in the
non-moving background.

end of update 10.