Update #11
I have:

++Added a "Win-Condition" Game State,

++Added 100 points counter for all enemies.

-+ I set up a  "Wave" system for the enemies and for the Invisible Win
Collider for the Win Condition for the player to meet the Win Collider.,

--Still the Need the art assets and i still need to fix the sound
effects.

how the "Invisible" WinBoxCollider works is that it has a spawner game
object to support the spawning of the WinBoxCollider, and it has set up
with a timing so that the WinBoxCollider can appear from the spawner.

I included a GameState Winning and i set up  and extra collision
detection of the "Invisible" WinBoxCollider, for everytime the player
collides with the WinBoxCollider, a function will activate the Winning
GameState, and once the transition is in effect it will stop the
Spawning of all enemies and the WinBoxCollider.